main-develop branch merge

+++ src/main/java/com/ssafy/enjoycamping/review/service/ReviewServiceImpl.java
@@ -1,61 +1,91 @@
 package com.ssafy.enjoycamping.review.service;
 
 
-import java.util.List;
+import java.io.IOException;
+import java.net.URL;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.*;
 
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Primary;
+import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest;
+import com.ssafy.enjoycamping.common.exception.ForbiddenException;
+import com.ssafy.enjoycamping.common.service.AsyncS3ImageService;
+import com.ssafy.enjoycamping.trip.camping.dto.CampingDto;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
-import com.ssafy.enjoycamping.common.exception.BadRequestException;
+import com.amazonaws.HttpMethod;
+import com.amazonaws.services.s3.AmazonS3;
 import com.ssafy.enjoycamping.common.exception.BaseException;
 import com.ssafy.enjoycamping.common.exception.NotFoundException;
-import com.ssafy.enjoycamping.common.exception.UnauthorizedException;
 import com.ssafy.enjoycamping.common.response.BaseResponseStatus;
 import com.ssafy.enjoycamping.common.util.PagingAndSorting;
 import com.ssafy.enjoycamping.review.dao.ReviewDao;
+import com.ssafy.enjoycamping.review.dao.ReviewImageDao;
 import com.ssafy.enjoycamping.review.dto.CreateReviewDto;
-import com.ssafy.enjoycamping.review.dto.CreateReviewDto.ResponseCreateReviewDto;
 import com.ssafy.enjoycamping.review.dto.ReviewDto;
 import com.ssafy.enjoycamping.review.dto.UpdateReviewDto;
 import com.ssafy.enjoycamping.review.entity.Review;
+import com.ssafy.enjoycamping.review.entity.ReviewImage;
 import com.ssafy.enjoycamping.trip.camping.dao.CampingDao;
 import com.ssafy.enjoycamping.trip.camping.entity.Camping;
 import com.ssafy.enjoycamping.user.dao.UserDao;
-import com.ssafy.enjoycamping.user.entity.User;
-import com.ssafy.enjoycamping.user.util.JwtProvider;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class ReviewServiceImpl implements ReviewService {
-	private ReviewDao reviewDao;
-	private UserDao userDao;
-	private CampingDao campingDao;
+	private final ReviewDao reviewDao;
+	private final ReviewImageDao reviewImageDao;
+	private final UserDao userDao;
+	private final CampingDao campingDao;
+	private final AmazonS3 amazonS3;
+	@Value("${aws.s3.bucket}")
+	private String bucket;
+	private final AsyncS3ImageService asyncS3ImageService;
 	
-	public ReviewServiceImpl(ReviewDao reviewDao, UserDao userDao, CampingDao campingDao) {
-		this.reviewDao = reviewDao;
-		this.userDao = userDao;
-		this.campingDao = campingDao;
-	}
-
+	@Transactional
 	@Override
-	public CreateReviewDto.ResponseCreateReviewDto createReview(CreateReviewDto.RequestCreateReviewDto request) {
-//		int id = JwtProvider.getUserId();
-//		// JWT로 User 불러오기 //access Token 만료됐는지 확인하기
-//		User user = userDao.selectActiveById(id)
-//				.orElseThrow(() -> new UnauthorizedException(BaseResponseStatus.INVALID_USER_JWT));
+	public CreateReviewDto.ResponseCreateReviewDto createReview(int userId, CreateReviewDto.RequestCreateReviewDto request) {
+		CampingDto camping = campingDao.selectById(request.getCampingId())
+				.orElseThrow(()->new NotFoundException(BaseResponseStatus.NOT_EXIST_CAMPING));
 		
-		//TODO: 이미지 들어왔을 때 이미지 테이블에 인서트하기
-		Review newReview = request.toEntity();
+		Review newReview = request.toEntity(camping, userId);
 		reviewDao.insert(newReview);
+
+		//이미지 맵핑 테이블에 이미지 URL저장
+		Set<String> imageUrls = Optional.ofNullable(request.getImageUrls()).orElse(Collections.emptySet());
+		if (!imageUrls.isEmpty()) reviewImageDao.insert(newReview.getId(), imageUrls);
+
 		return CreateReviewDto.ResponseCreateReviewDto.builder()
 				.id(newReview.getId())
 				.build();
 	}
 	
+	@Override
+	public URL createImageUrl(String fileName, String contentType) throws IOException {
+		// S3 객체 키 (파일 이름)
+		String objectKey = "uploads/" + fileName;
+		Date expireTime = Date.from(Instant.now().plus(1,ChronoUnit.MINUTES));
+
+		GeneratePresignedUrlRequest request = new GeneratePresignedUrlRequest(bucket, objectKey)
+				.withMethod(HttpMethod.PUT)
+				.withContentType(contentType)
+				.withExpiration(expireTime);
+
+		// 'x-amz-acl' 헤더를 'public-read'로 설정하여 퍼블릭 읽기 권한 부여
+		request.addRequestParameter("x-amz-acl", "public-read");
+
+		//preSignedUrl 발급
+		return amazonS3.generatePresignedUrl(request);
+	}
+
 	@Override
 	public ReviewDto getReview(int id) throws BaseException {
 		Review review = reviewDao.selectById(id)
@@ -63,21 +93,56 @@ public class ReviewServiceImpl implements ReviewService {
 		return ReviewDto.fromEntity(review);
 	}
 
+	@Transactional
 	@Override
-	public void deleteReview(int id) {
-		//TODO: 로그인 유저와 작성자 확인 후 맞으면 삭제하는 로직
+	public void deleteReview(int userId, int id) {
 		Review review = reviewDao.selectById(id)
 				.orElseThrow(() -> new NotFoundException(BaseResponseStatus.NOT_EXIST_REVIEW));
+
+		if(review.getWriterId() != userId) throw new ForbiddenException(BaseResponseStatus.INVALID_USER_JWT);
+
+		Set<String> ImageUrlsToDelete = reviewImageDao.selectImageUrlsByReviewId(id);
+		if(!ImageUrlsToDelete.isEmpty()) {
+			reviewImageDao.delete(ImageUrlsToDelete);
+			asyncS3ImageService.deleteImagesFromS3(ImageUrlsToDelete);
+		}
 		reviewDao.delete(id);
+		log.info("삭제 완료");
 	}
 
+	@Transactional
 	@Override
-	public ReviewDto updateReview(UpdateReviewDto.RequestUpdateReviewDto request, int id) {
-		//TODO: 로그인 유저와 작성자 확인 후 맞으면 업데이트하는 로직
+	public ReviewDto updateReview(int userId, UpdateReviewDto.RequestUpdateReviewDto request, int id) {
 		Review review = reviewDao.selectById(id)
 				.orElseThrow(() -> new NotFoundException(BaseResponseStatus.NOT_EXIST_REVIEW));
+
+		if(review.getWriterId() != userId) throw new ForbiddenException(BaseResponseStatus.INVALID_USER_JWT);
 		
-		request.updateReview(review);
+		//새로운 이미지 URL
+		Set<String> newReviewImages = Optional.ofNullable(request.getImageUrls()).orElse(new HashSet<>());
+		//기존 이미지 URL
+		Set<String> originReviewImages = reviewImageDao.selectImageUrlsByReviewId(id);
+		
+		//추가된 이미지
+		Set<String>imagesToInsert = new HashSet<>(newReviewImages);
+		imagesToInsert.removeAll(originReviewImages);
+		
+		//삭제될 이미지
+		Set<String>imagesToDelete = new HashSet<>(originReviewImages);
+		imagesToDelete.removeAll(newReviewImages);
+		
+		//추가된 이미지가 있다면 추가
+		if(!imagesToInsert.isEmpty()) {
+			reviewImageDao.insert(id, imagesToInsert);
+		}
+		//삭제될 이미지가 있다면 삭제
+		if(!imagesToDelete.isEmpty()) {
+			//S3 이미지 삭제
+			reviewImageDao.delete(imagesToDelete);
+			asyncS3ImageService.deleteImagesFromS3(imagesToDelete);
+		}
+		
+		review.updateReview(request);
 		reviewDao.update(review);
 		return ReviewDto.fromEntity(review);
 	}
@@ -93,7 +158,7 @@ public class ReviewServiceImpl implements ReviewService {
 	
 	@Override
 	public List<ReviewDto> getReviewsByCampingId(int campingId) throws BaseException{
-		Camping camping = campingDao.selectById(campingId)
+		CampingDto camping = campingDao.selectById(campingId)
 				.orElseThrow(()-> new NotFoundException(BaseResponseStatus.NOT_EXIST_CAMPING));
 		
 		List<ReviewDto> reviews = reviewDao.selectByCampingId(camping.getId())
@@ -123,8 +188,5 @@ public class ReviewServiceImpl implements ReviewService {
 				.map(ReviewDto::fromEntity)
 				.toList();
 	}
-	
-	
-
 
 }
